+++ sklearn/svm/tests/test_svm.py
@@ -255,9 +255,9 @@ def test_weight():
     """
     Test class weights
     """
-    clf = svm.SVC()
+    clf = svm.SVC(class_weight={1: 0.1})
     # we give a small weights to class 1
-    clf.fit(X, Y, {1: 0.1})
+    clf.fit(X, Y)
     # so all predicted values belong to class 2
     assert_array_almost_equal(clf.predict(X), [2] * 6)
 
@@ -265,7 +265,8 @@ def test_weight():
                                  weights=[0.833, 0.167], random_state=0)
 
     for clf in (linear_model.LogisticRegression(), svm.LinearSVC(), svm.SVC()):
-        clf.fit(X_[: 180], y_[: 180], class_weight={0: 5})
+        clf.set_params(class_weight={0: 5})
+        clf.fit(X_[: 180], y_[: 180])
         y_pred = clf.predict(X_[180:])
         assert np.sum(y_pred == y_[180:]) >= 11
 
@@ -299,10 +300,9 @@ def test_auto_weight():
     for clf in (svm.SVC(kernel='linear'),
             svm.LinearSVC(), LogisticRegression()):
         # check that score is better when class='auto' is set.
-        y_pred = clf.fit(X[unbalanced], y[unbalanced],
-                         class_weight={}).predict(X)
-        y_pred_balanced = clf.fit(X[unbalanced], y[unbalanced],
-                                  class_weight='auto').predict(X)
+        y_pred = clf.fit(X[unbalanced], y[unbalanced]).predict(X)
+        clf.set_params(class_weight='auto')
+        y_pred_balanced = clf.fit(X[unbalanced], y[unbalanced],).predict(X)
         assert metrics.f1_score(y, y_pred) <= \
                 metrics.f1_score(y, y_pred_balanced)